ST-294:Refactor of test packages
